Fix lint issue
