+++ Model/Config.php
@@ -21,7 +21,6 @@ use Magento\Store\Model\ScopeInterface;
 use Magento\Store\Model\Store;
 use Magento\Framework\Serialize\Serializer\Json;
 use Magento\Framework\App\Config\Storage\WriterInterface;
-use Magento\Framework\Composer\ComposerInformation;
 
 /**
  * @SuppressWarnings(PHPMD.ExcessiveClassComplexity)
@@ -121,7 +120,6 @@ class Config
      * @param State $state
      * @param WriterInterface $configWriter
      * @param TypeListInterface $cacheTypeList
-     * @param ComposerInformation $composerInformation
      * @throws LocalizedException
      */
     public function __construct(
@@ -130,8 +128,7 @@ class Config
         RequestInterface $request,
         State $state,
         WriterInterface $configWriter,
-        TypeListInterface $cacheTypeList,
-        private readonly ComposerInformation $composerInformation
+        TypeListInterface $cacheTypeList
     ) {
         $this->config = $config;
         $this->jsonSerializer = $jsonSerializer;
@@ -640,22 +637,5 @@ class Config
     public function getProductCardLifetime(): int
     {
         return (int) $this->config->getValue(self::PRODUCT_CARD_LIFETIME_XML_PATH, ScopeInterface::SCOPE_STORE);
-    }
-
-    /**
-     * @return string
-     */
-    public function getModuleVersion(): string
-    {
-        $installedPackages = $this->composerInformation
-            ->getInstalledMagentoPackages();
-        if (!isset($installedPackages['tweakwise/magento2-tweakwise']['version'])) {
-            // This should never be the case
-            return '';
-        }
-
-        $version = $installedPackages['tweakwise/magento2-tweakwise']['version'];
-
-        return sprintf('Magento2Tweakwise %s', $version);
     }
 }
